requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.9.0

Pillow==12.1.1
openai==1.92.0
PyMySQL==1.1.2
